分离查询相关代码， 避免误修改导致接口不兼容
