+++ plugin/dapp/trade/executor/query.go
@@ -24,11 +24,6 @@ import (
 // 由于现价买/卖是没有orderID的， 用txhash 代替作为key
 // key 有两种 orderID， txhash (0xAAAAAAAAAAAAAAA)
 
-// 根据token 分页显示未完成成交卖单
-func (t *trade) Query_GetTokenSellOrderByStatus(req *pty.ReqTokenSellOrder) (types.Message, error) {
-	return t.GetTokenSellOrderByStatus(req, req.Status)
-}
-
 // 根据token 分页显示未完成成交买单
 func (t *trade) Query_GetTokenBuyOrderByStatus(req *pty.ReqTokenBuyOrder) (types.Message, error) {
 	if req.Status == 0 {
@@ -181,36 +176,6 @@ func (t *trade) GetOnesBuyOrdersWithStatus(req *pty.ReqAddrAssets) (types.Messag
 		replys.Orders = append(replys.Orders, reply)
 	}
 
-	return &replys, nil
-}
-
-func (t *trade) GetTokenSellOrderByStatus(req *pty.ReqTokenSellOrder, status int32) (types.Message, error) {
-	if req.Count <= 0 || (req.Direction != 1 && req.Direction != 0) {
-		return nil, types.ErrInvalidParam
-	}
-
-	fromKey := []byte("")
-	if len(req.FromKey) != 0 {
-		sell := t.replyReplySellOrderfromID([]byte(req.FromKey))
-		if sell == nil {
-			tradelog.Error("GetTokenSellOrderByStatus", "key not exist", req.FromKey)
-			return nil, types.ErrInvalidParam
-		}
-		fromKey = calcTokensSellOrderKeyStatus(sell.TokenSymbol, sell.Status,
-			calcPriceOfToken(sell.PricePerBoardlot, sell.AmountPerBoardlot), sell.Owner, sell.Key)
-	}
-	values, err := t.GetLocalDB().List(calcTokensSellOrderPrefixStatus(req.TokenSymbol, status), fromKey, req.Count, req.Direction)
-	if err != nil {
-		return nil, err
-	}
-	var replys pty.ReplyTradeOrders
-	for _, key := range values {
-		reply := t.loadOrderFromKey(key)
-		if reply == nil {
-			continue
-		}
-		replys.Orders = append(replys.Orders, reply)
-	}
 	return &replys, nil
 }
