+++ plugin/dapp/trade/executor/query_v1.go
@@ -1,0 +1,41 @@
+package executor
+
+import (
+	"github.com/33cn/chain33/types"
+	pty "github.com/33cn/plugin/plugin/dapp/trade/types"
+)
+
+// 根据token 分页显示未完成成交卖单
+func (t *trade) Query_GetTokenSellOrderByStatus(req *pty.ReqTokenSellOrder) (types.Message, error) {
+	return t.GetTokenSellOrderByStatus(req, req.Status)
+}
+
+func (t *trade) GetTokenSellOrderByStatus(req *pty.ReqTokenSellOrder, status int32) (types.Message, error) {
+	if req.Count <= 0 || (req.Direction != 1 && req.Direction != 0) {
+		return nil, types.ErrInvalidParam
+	}
+
+	fromKey := []byte("")
+	if len(req.FromKey) != 0 {
+		sell := t.replyReplySellOrderfromID([]byte(req.FromKey))
+		if sell == nil {
+			tradelog.Error("GetTokenSellOrderByStatus", "key not exist", req.FromKey)
+			return nil, types.ErrInvalidParam
+		}
+		fromKey = calcTokensSellOrderKeyStatus(sell.TokenSymbol, sell.Status,
+			calcPriceOfToken(sell.PricePerBoardlot, sell.AmountPerBoardlot), sell.Owner, sell.Key)
+	}
+	values, err := t.GetLocalDB().List(calcTokensSellOrderPrefixStatus(req.TokenSymbol, status), fromKey, req.Count, req.Direction)
+	if err != nil {
+		return nil, err
+	}
+	var replys pty.ReplyTradeOrders
+	for _, key := range values {
+		reply := t.loadOrderFromKey(key)
+		if reply == nil {
+			continue
+		}
+		replys.Orders = append(replys.Orders, reply)
+	}
+	return &replys, nil
+}
